basic navigation implimentation
