+++ Hproj-mobile/MProj/index.js
@@ -1,7 +1,7 @@
 /** @format */
 
 import {AppRegistry} from 'react-native';
-import Root from './src';
+import AppContainer from './src/index'
 import {name as appName} from './app.json';
 
-AppRegistry.registerComponent(appName, () => Root);
+AppRegistry.registerComponent(appName, () => AppContainer);
