+++ Hproj-mobile/MProj/package.json
@@ -8,7 +8,9 @@
   },
   "dependencies": {
     "react": "16.6.1",
-    "react-native": "0.57.5"
+    "react-native": "0.57.5",
+    "react-native-gesture-handler": "^1.0.10",
+    "react-navigation": "^3.0.7"
   },
   "devDependencies": {
     "babel-jest": "23.6.0",
